chore: Slight refactor of types/interface
