+++ src/app/zaken/ZaakInterface.ts
@@ -1,28 +1,34 @@
-export interface ZaakSummary {
-  identifier: string;
-  internal_id: string;
-  registratiedatum: Date;
-  verwachtte_einddatum?: Date;
-  uiterlijke_einddatum?: Date;
-  einddatum?: Date | null | undefined;
-  zaak_type: string;
-  status: string | null;
-  resultaat?: string | null;
-}
+import { z } from 'zod';
 
-export interface SingleZaak {
-  identifier: string;
-  internal_id: string;
-  registratiedatum?: Date;
-  verwachtte_einddatum?: Date;
-  uiterlijke_einddatum?: Date;
-  einddatum?: Date;
-  zaak_type: string;
-  status: string;
-  status_list?: any[];
-  resultaat?: string;
-  documenten?: any[];
-  taken?: any[];
-  behandelaars?: string[];
-  type: 'case' | 'submission';
-}
+export const ZaakSummarySchema = z.object({
+  identifier: z.string(),
+  internal_id: z.string(),
+  registratiedatum: z.coerce.date(),
+  verwachtte_einddatum: z.coerce.date().optional(),
+  uiterlijke_einddatum: z.coerce.date().optional(),
+  einddatum: z.coerce.date().optional().nullable(),
+  zaak_type: z.string(),
+  status: z.string().nullable(),
+  resultaat: z.string().optional().nullable(),
+}).passthrough();
+export const ZaakSummariesSchema = z.array(ZaakSummarySchema);
+
+export type ZaakSummary = z.infer<typeof ZaakSummarySchema>;
+
+export const singleZaakSchema = z.object({
+  identifier: z.string(),
+  internal_id: z.string(),
+  registratiedatum: z.date().optional(),
+  verwachtte_einddatum: z.date().optional(),
+  uiterlijke_einddatum: z.date().optional(),
+  einddatum: z.date().optional(),
+  zaak_type: z.string(),
+  status: z.string().optional(),
+  status_list: z.array(z.any()).optional(),
+  resultaat: z.string().optional(),
+  documenten: z.array(z.any()).optional(),
+  taken: z.array(z.any()).optional(),
+  behandelaars: z.array(z.string()).optional(),
+  type: z.enum(['case', 'submission']),
+});
+export type SingleZaak = z.infer<typeof singleZaakSchema>;
